gql: move hasura logic into new client for context (#3934)
