+++ backend/native/backpack-api/src/routes/graphql/clients/index.ts
@@ -1,3 +1,4 @@
 export { CoinGecko } from "./coingecko";
+export { Hasura } from "./hasura";
 export { Helius } from "./helius";
 export { Tensor } from "./tensor";
